+++ resources/js/Pages/Entries/Index.vue
@@ -33,12 +33,24 @@
                     </div>
                 </td>
             </template>
+
+            <template v-slot:footer>
+                <div class="flex items-center justify-center font-bold">
+                    <f-button v-if="items.meta.currentPage > 1" :href="items.meta.firstPageUrl" color="brand" variant="outline" class="mx-1"><f-icon name="step-backward" /></f-button>
+                    <f-button v-if="items.meta.previousPageUrl" :href="items.meta.previousPageUrl" color="brand" variant="outline" class="mx-1"><f-icon name="angle-left" /></f-button>
+                    <span v-if="hasMoreBefore" class="mx-1 px-1 text-gray-700"><f-icon name="ellipsis-h" /></span>
+                    <f-button v-for="page in items.meta.pageUrls" :key="page.number" :href="page.url" color="brand" :variant="page.isCurrent ? 'solid' : 'outline'" class="mx-1">{{ page.number }}</f-button>
+                    <span v-if="hasMoreAfter" class="mx-1 px-1 text-gray-700"><f-icon name="ellipsis-h" /></span>
+                    <f-button v-if="items.meta.nextPageUrl" :href="items.meta.nextPageUrl" color="brand" variant="outline" class="mx-1"><f-icon name="angle-right" /></f-button>
+                    <f-button v-if="items.meta.currentPage < items.meta.lastPage" :href="items.meta.lastPageUrl" color="brand" variant="outline" class="mx-1"><f-icon name="step-forward" /></f-button>
+                </div>
+            </template>
         </f-table-card>
     </f-the-app-layout>
 </template>
 
 <script>
-
+import map from 'lodash/map'
 import ListSchema from '../../Support/ListSchema'
 
 export default {
@@ -55,6 +67,24 @@ export default {
         return {
             isPerformingActionFor: {},
             listSchema: new ListSchema(this.items.meta.entry_type.schema),
+        }
+    },
+
+    computed: {
+        hasMoreBefore () {
+            if (this.items.meta.pageUrls.length === 0) {
+                return false;
+            }
+
+            return this.items.meta.pageUrls[0].number > 1
+        },
+
+        hasMoreAfter () {
+            if (this.items.meta.pageUrls.length === 0) {
+                return false;
+            }
+
+            return this.items.meta.pageUrls[this.items.meta.pageUrls.length - 1].number < this.items.meta.lastPage
         }
     },