+++ src/EntryTypes/EntryType.php
@@ -16,6 +16,7 @@ use CbtechLtd\Fastlane\Support\Schema\Fields\Contracts\WriteValue;
 use CbtechLtd\Fastlane\Support\Schema\Fields\Contracts\WithRules;
 use CbtechLtd\JsonApiTransformer\ApiResources\ResourceType;
 use Illuminate\Contracts\Auth\Access\Gate;
+use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
@@ -154,7 +155,7 @@ abstract class EntryType implements EntryTypeContract
         return true;
     }
 
-    public function getItems(): Collection
+    public function getItems(): LengthAwarePaginator
     {
         $this->gate->authorize('list', $this->model());
 
@@ -166,9 +167,11 @@ abstract class EntryType implements EntryTypeContract
 
         $this->queryItems($query);
 
-        return Collection::make($query->get())->map(function (Model $model) {
-            return $this->newInstance($model);
-        });
+        $pagination = $query->paginate($this->getItemsPerPage());
+
+        $pagination->getCollection()->transform(fn(Model $model) => $this->newInstance($model));
+
+        return $pagination;
     }
 
     public function findItem(string $hashid): EntryInstanceContract
@@ -204,7 +207,7 @@ abstract class EntryType implements EntryTypeContract
             ->all();
 
         $data = Validator::make($request->all(), $rules)->validated();
-        
+
         // Pass the validated date through all fields, call hooks before saving,
         // then call more hooks after model has been saved.
         $this->hydrateFields($entryInstance, $fields, $data);
@@ -354,5 +357,10 @@ abstract class EntryType implements EntryTypeContract
                     }
                 }
             });
+    }
+
+    protected function getItemsPerPage(): int
+    {
+        return config('fastlane.control_panel.pagination_per_page');
     }
 }